pasando int a float64

+++ microservice/internal/adapter/handler/request/coupon.go
@@ -1,8 +1,8 @@
 package request
 
 type RequestPaymentMicroservice struct {
-	Amount         int    `json:"amount"`
-	Description    string `json:"description"`
-	SuccesResponse string `json:"succesResponse"`
-	FailedResponse string `json:"failedResponse"`
+	Amount         float64 `json:"amount"`
+	Description    string  `json:"description"`
+	SuccesResponse string  `json:"succesResponse"`
+	FailedResponse string  `json:"failedResponse"`
 }
