+++ microservice/internal/adapter/handler/request/user.go
@@ -13,11 +13,11 @@ type LoginUserRequest struct {
 }
 
 type CreateGiftCardRequest struct {
-	Title         string `json:"title" binding:"required"`
-	Email         string `json:"email" binding:"required"`
-	Description   string `json:"description" binding:"required"`
-	AmountCoupons int    `json:"amount" binding:"required"`
-	PriceCoupons  int    `json:"price" binding:"required"`
+	Title         string  `json:"title" binding:"required"`
+	Email         string  `json:"email" binding:"required"`
+	Description   string  `json:"description" binding:"required"`
+	AmountCoupons int     `json:"amount" binding:"required"`
+	PriceCoupons  float64 `json:"price" binding:"required"`
 }
 
 type InsertCodeRequest struct {
